requirements.txt
behave==1.2.6
behave_django==1.1.0
coverage==3.7.1
django-selenium-login==1.0.2
flake8==3.6.0
flake8-import-order==0.18
hypothesis==2.0.0
psycopg2==2.7.4
selenium==3.14.1
